[Feat] 계정 삭제, 비밀번호 변경, 코드를 받아 비밀번호 재설정 기능 API 구현

+++ src/main/java/com/example/shoppingmallserver/service/AuthService.java
@@ -52,5 +52,5 @@ public interface AuthService {
      *
      * 사용자의 고유 식별자를 기반으로 해당 사용자 계정을 삭제합니다.
      */
-    void deleteAccount(int identifier);
+    void deleteAccount(Long identifier);
 }
